requirements.txt
fastapi
uvicorn
python-dotenv
requests
langchain
langchain-openai
langchain-community
langgraph
ddgs